Add a tile selection frame
This frame doesn't quite do anything yet. I spent a little while (30-60
minutes) trying to figure out the pack geometry manager such that the
frames fit side by side and the canvas displayed properly.

The selection frame is 'packed in' first, and sent to the left side.
There is an x padding added since otherwise only one of the frames would
show up.

The canvas widget was also modified so that a) it was easier to see in
relation to the other frame and b) scaled more dynamically with the
parent frame.